requirements.txt
click
rich
python-dotenv
inquirer
yaspin
numpy
soundfile
